refactor(petform): refactored most state and functions into form context
